wallet_project/requirements.txt
Django==5.1.0
djangorestframework==3.14.0
psycopg2==2.9.3
